requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
docker>=7.0.0
requests>=2.31.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
pydantic>=2.0.0
